Update visualization.html
Point to note: Most of the features being added are only as per the list in the text file. The functions themselves would require some modification before any assimilation with the real data can be made. I'm currently leaving certain if conditionals (for currently and prev. speaking ) to be as they might come in handy later on.